fix(dictionary,refcatalog): address code review findings
- Fix HardDeleteOld return type: int → int64 to match postgres repo
- Fix refcatalog clampLimit: default 0 → 20 (not 1) per spec
- Add Example.Translation length validation (max 2000)
- Replace custom itoa with strconv.Itoa

+++ backend_v4/internal/service/dictionary/service.go
@@ -29,7 +29,7 @@ type entryRepo interface {
 	UpdateNotes(ctx context.Context, userID, entryID uuid.UUID, notes *string) (*domain.Entry, error)
 	SoftDelete(ctx context.Context, userID, entryID uuid.UUID) error
 	Restore(ctx context.Context, userID, entryID uuid.UUID) (*domain.Entry, error)
-	HardDeleteOld(ctx context.Context, threshold time.Time) (int, error)
+	HardDeleteOld(ctx context.Context, threshold time.Time) (int64, error)
 }
 
 type senseRepo interface {
